+++ core/main_method/entities.py
@@ -20,6 +20,7 @@ from core.entities import (
     EditPointDialog,
     EditLinkDialog,
 )
+from core.libs import expr_solving
 from core.graphics import edges_view
 from core.io import (
     AddTable,
@@ -315,7 +316,7 @@ def clonePoint(self):
     self.CommandStack.endMacro()
 
 
-def setFreemoved(self,
+def setFreemove(self,
     coords: Tuple[Tuple[int, Tuple[float, float, float]]]
 ):
     """Free move function."""
@@ -335,6 +336,45 @@ def setFreemoved(self,
             args
         ))
     self.CommandStack.endMacro()
+
+
+def adjustLinkage(self, value: int):
+    """Preview the free move result."""
+    vpoints = self.EntitiesPoint.dataTuple()
+    mapping = {n: 'P{}'.format(n) for n in range(len(vpoints))}
+    mapping[self.linkage_freemode_linkname.text()] = float(value)
+    try:
+        result = expr_solving(
+            self.getTriangle(),
+            mapping,
+            vpoints,
+            [v[-1] for v in self.InputsWidget.getInputsVariables()]
+        )
+    except Exception:
+        pass
+    else:
+        self.MainCanvas.adjustLinkage(result)
+        if not self.linkage_freemode_slider.isSliderDown():
+            self.MainCanvas.emit_freemove_all()
+
+
+def setLinkageFreemove(self, enable: bool):
+    """Free move function for linkage length."""
+    self.linkage_freemode_widget.setEnabled(enable)
+    self.linkage_freemode_linkname.clear()
+    if not enable:
+        return
+    item = self.EntitiesExpr.currentItem()
+    if not item:
+        return
+    name, value = item.text().split(':')
+    self.linkage_freemode_linkname.setText(name)
+    try:
+        self.linkage_freemode_slider.valueChanged.disconnect(self.adjustLinkage)
+    except TypeError:
+        pass
+    self.linkage_freemode_slider.setValue(float(value))
+    self.linkage_freemode_slider.valueChanged.connect(self.adjustLinkage)
 
 
 def on_action_New_Link_triggered(self):
@@ -479,7 +519,7 @@ def on_action_Delete_Link_triggered(self):
 
 def setCoordsAsCurrent(self):
     """Update points position as current coordinate."""
-    self.setFreemoved(tuple(
+    self.setFreemove(tuple(
         (row, self.EntitiesPoint.currentPosition(row)[0])
         for row in range(self.EntitiesPoint.rowCount())
     ))