+++ core/widgets/main_canvas_method.py
@@ -139,7 +139,7 @@ def _drawFrame(self):
 
 def _drawPoint(self, i: int, vpoint: VPoint):
     """Draw a point."""
-    if vpoint.type==1 or vpoint.type==2:
+    if vpoint.type in (1, 2):
         #Draw slider
         silder_points = vpoint.c
         for j, (cx, cy) in enumerate(silder_points):
@@ -149,7 +149,7 @@ def _drawPoint(self, i: int, vpoint: VPoint):
                 grounded = vpoint.links[j] == 'ground'
             if vpoint.type == 1:
                 if j == 0:
-                    self._BaseCanvas__drawPoint(
+                    self.drawPoint(
                         i, cx, cy,
                         grounded,
                         vpoint.color
@@ -165,21 +165,21 @@ def _drawPoint(self, i: int, vpoint: VPoint):
                     ))
             elif vpoint.type == 2:
                 if j == 0:
-                    self._BaseCanvas__drawPoint(
+                    self.drawPoint(
                         i, cx, cy,
                         grounded,
                         vpoint.color
                     )
                 else:
                     #Turn off point mark.
-                    showPointMark = self.showPointMark
-                    self.showPointMark = False
-                    self._BaseCanvas__drawPoint(
+                    show_point_mark = self.show_point_mark
+                    self.show_point_mark = False
+                    self.drawPoint(
                         i, cx, cy,
                         grounded,
                         vpoint.color
                     )
-                    self.showPointMark = showPointMark
+                    self.show_point_mark = show_point_mark
         pen = QPen(vpoint.color.darker())
         pen.setWidth(2)
         self.painter.setPen(pen)
@@ -196,13 +196,13 @@ def _drawPoint(self, i: int, vpoint: VPoint):
                 QPointF(p_right[0] * self.zoom, p_right[1] * -self.zoom)
             )
     else:
-        self._BaseCanvas__drawPoint(
+        self.drawPoint(
             i, vpoint.cx, vpoint.cy,
             vpoint.grounded(),
             vpoint.color
         )
     #For selects function.
-    if (self.selectionMode == 0) and (i in self.selections):
+    if (self.select_mode == 0) and (i in self.selections):
         pen = QPen(QColor(161, 16, 239))
         pen.setWidth(3)
         self.painter.setPen(pen)
@@ -240,14 +240,14 @@ def _drawLink(self, vlink: VLink):
     #Rearrange: Put the nearest point to the next position.
     qpoints = convex_hull(points, as_qpoint=True)
     if (
-        (self.selectionMode == 1) and
+        (self.select_mode == 1) and
         (self.vlinks.index(vlink) in self.selections)
     ):
-        pen.setWidth(self.linkWidth + 6)
+        pen.setWidth(self.link_width + 6)
         pen.setColor(QColor(161, 16, 239))
         self.painter.setPen(pen)
         self.painter.drawPolygon(*qpoints)
-    pen.setWidth(self.linkWidth)
+    pen.setWidth(self.link_width)
     pen.setColor(vlink.color)
     self.painter.setPen(pen)
     brush = QColor(226, 219, 190)
@@ -255,7 +255,7 @@ def _drawLink(self, vlink: VLink):
     self.painter.setBrush(brush)
     self.painter.drawPolygon(*qpoints)
     self.painter.setBrush(Qt.NoBrush)
-    if not self.showPointMark:
+    if not self.show_point_mark:
         return
     pen.setColor(Qt.darkGray)
     self.painter.setPen(pen)
@@ -272,23 +272,17 @@ def _drawLink(self, vlink: VLink):
 def _drawPath(self):
     """Draw paths. Recording first."""
     pen = QPen()
-    if self.autoPathShow and self.rightInput():
+    if self.auto_path and self.rightInput():
         """Replace to auto preview path."""
-        exprs = self.getTriangle(self.vpoints)
+        self.exprs = self.getTriangle(self.vpoints)
         self.Path.path = expr_path(
-            exprs,
+            self.exprs,
             {n: 'P{}'.format(n) for n in range(len(self.vpoints))},
             self.vpoints,
             self.pathInterval()
         )
-        if self.selectionMode == 2:
-            for expr in exprs:
-                self._BaseCanvas__drawSolution(
-                    expr[0],
-                    expr[1:-1],
-                    expr[-1],
-                    self.vpoints
-                )
+    else:
+        self.exprs = []
     if hasattr(self, 'path_record'):
         paths = self.path_record
     else:
@@ -305,18 +299,17 @@ def _drawPath(self):
         except:
             color = colorQt('Green')
         pen.setColor(color)
-        pen.setWidth(self.pathWidth)
+        pen.setWidth(self.path_width)
         self.painter.setPen(pen)
         if self.Path.curve:
-            self._BaseCanvas__drawCurve(path)
+            self.drawCurve(path)
         else:
-            self._BaseCanvas__drawDot(path)
+            self.drawDot(path)
 
 
 def _drawSlvsRanges(self):
     """Draw solving range."""
     pen = QPen()
-    self.painter.setFont(QFont("Arial", self.fontSize + 5))
     pen.setWidth(5)
     for i, (tag, rect) in enumerate(self.ranges.items()):
         range_color = QColor(colorNum(i+1))
@@ -343,10 +336,19 @@ def _drawSlvsRanges(self):
         self.painter.setBrush(Qt.NoBrush)
 
 
+def _emit_freemove(self, targets: List[int]):
+    """Emit free move targets to edit."""
+    self.freemoved.emit(tuple((num, (
+        self.vpoints[num].cx,
+        self.vpoints[num].cy,
+        self.vpoints[num].angle,
+    )) for num in targets))
+
+
 def _select_func(self, *, rect: bool = False):
     """Select function."""
     self.selector.selection_rect.clear()
-    if self.selectionMode == 0:
+    if self.select_mode == 0:
         
         def catch(x: float, y: float) -> bool:
             """Detection function for points."""
@@ -359,7 +361,8 @@ def _select_func(self, *, rect: bool = False):
             if catch(vpoint.cx * self.zoom, vpoint.cy * -self.zoom):
                 if i not in self.selector.selection_rect:
                     self.selector.selection_rect.append(i)
-    elif self.selectionMode == 1:
+        
+    elif self.select_mode == 1:
         
         def catch(vlink: VLink) -> bool:
             """Detection function for links.
@@ -371,13 +374,18 @@ def _select_func(self, *, rect: bool = False):
             if len(points) > 2:
                 polygon = QPolygonF(convex_hull(points, as_qpoint=True))
             else:
-                points_up = [(x + self.sr, y + self.sr) for x, y in points]
-                points_down = [(x - self.sr, y - self.sr) for x, y in points]
-                polygon = QPolygonF(convex_hull(points_up + points_down, as_qpoint=True))
+                polygon = QPolygonF(convex_hull(
+                    [(x + self.sr, y + self.sr) for x, y in points] +
+                    [(x - self.sr, y - self.sr) for x, y in points],
+                    as_qpoint=True
+                ))
             if rect:
                 return polygon.intersects(QPolygonF(self.selector.toQRect()))
             else:
-                return polygon.containsPoint(QPointF(self.selector.x, self.selector.y), Qt.WindingFill)
+                return polygon.containsPoint(
+                    QPointF(self.selector.x, self.selector.y),
+                    Qt.WindingFill
+                )
         
         for i, vlink in enumerate(self.vlinks):
             if i == 0:
@@ -385,11 +393,35 @@ def _select_func(self, *, rect: bool = False):
             if catch(vlink):
                 if i not in self.selector.selection_rect:
                     self.selector.selection_rect.append(i)
+        
+    elif self.select_mode == 2:
+        
+        def catch(expr: str) -> bool:
+            """Detection function for solution polygons."""
+            points, _ = self.solutionPolygon(
+                expr[0],
+                expr[1:-1],
+                expr[-1],
+                self.vpoints
+            )
+            polygon = QPolygonF([QPointF(x, -y) * self.zoom for x, y in points])
+            if rect:
+                return polygon.intersects(QPolygonF(self.selector.toQRect()))
+            else:
+                return polygon.containsPoint(
+                    QPointF(self.selector.x, self.selector.y),
+                    Qt.WindingFill
+                )
+        
+        for i, expr in enumerate(self.exprs):
+            if catch(expr):
+                if i not in self.selector.selection_rect:
+                    self.selector.selection_rect.append(i)
 
 
-def _snap(self, num: float, isZoom: bool = True) -> float:
+def _snap(self, num: float, *, is_zoom: bool = True) -> float:
     """Close to a multiple of coefficient."""
-    snap_val = self.snap * self.zoom if isZoom else self.snap
+    snap_val = self.snap * self.zoom if is_zoom else self.snap
     if not snap_val:
         return num
     times = num // snap_val
@@ -435,8 +467,8 @@ def _zoomToFitLimit(self) -> Tuple[float, float, float, float]:
         if vpoint.cy > y_top:
             y_top = vpoint.cy
     #Solving paths
-    if self.showTargetPath:
-        for path in self.targetPath.values():
+    if self.show_target_path:
+        for path in self.target_path.values():
             for x, y in path:
                 if x < x_right:
                     x_right = x
@@ -463,6 +495,11 @@ def _zoomToFitLimit(self) -> Tuple[float, float, float, float]:
     return x_right, x_left, y_top, y_bottom
 
 
+def emit_freemove_all(self):
+    """Edit all points to edit."""
+    _emit_freemove(self, list(range(len(self.vpoints))))
+
+
 def paintEvent(self, event):
     """Drawing functions."""
     width = self.width()
@@ -475,7 +512,6 @@ def paintEvent(self, event):
         self.oy += (height - self.height_old) / 2
     #'self' is the instance of 'DynamicCanvas'.
     BaseCanvas.paintEvent(self, event)
-    self.painter.setFont(QFont('Arial', self.fontSize))
     #Draw links except ground.
     for vlink in self.vlinks[1:]:
         _drawLink(self, vlink)
@@ -483,17 +519,38 @@ def paintEvent(self, event):
     if self.Path.show != -2:
         _drawPath(self)
     #Draw solving path.
-    if self.showTargetPath:
+    if self.show_target_path:
+        self.painter.setFont(QFont("Arial", self.font_size + 5))
         _drawSlvsRanges(self)
-        self._BaseCanvas__drawTargetPath()
+        self.drawTargetPath()
+        self.painter.setFont(QFont("Arial", self.font_size))
     #Draw points.
     for i, vpoint in enumerate(self.vpoints):
         _drawPoint(self, i, vpoint)
+    #Draw solutions.
+    if self.select_mode == 2:
+        for i, expr in enumerate(self.exprs):
+            args = [
+                expr[0],
+                expr[1:-1],
+                expr[-1],
+                self.vpoints
+            ]
+            self.drawSolution(*args)
+            if i in self.selections:
+                pos, _ = self.solutionPolygon(*args)
+                pen = QPen()
+                pen.setWidth(self.link_width + 3)
+                pen.setColor(QColor(161, 16, 239))
+                self.painter.setPen(pen)
+                self.painter.drawPolygon(*[
+                    QPointF(x, -y) * self.zoom for x, y in pos
+                ])
     #Draw a colored frame for free move mode.
     if self.freemove != FreeMode.NoFreeMove:
         pen = QPen()
         if self.freemove == FreeMode.Translate:
-            pen.setColor(QColor(161, 105, 229))
+            pen.setColor(QColor(161, 16, 229))
         elif self.freemove == FreeMode.Rotate:
             pen.setColor(QColor(219, 162, 6))
         elif self.freemove == FreeMode.Reflect:
@@ -511,16 +568,6 @@ def paintEvent(self, event):
     #Record the widget size.
     self.width_old = width
     self.height_old = height
-
-
-def getRecordPath(self) -> Tuple[Tuple[Tuple[float, float]]]:
-    """Return paths."""
-    path = tuple(
-        tuple(path) if (len(set(path)) > 1) else ()
-        for path in self.path_record
-    )
-    del self.path_record
-    return path
 
 
 def mousePressEvent(self, event):
@@ -587,12 +634,8 @@ def mouseReleaseEvent(self, event):
             ):
                 self.noselected.emit()
         #Edit point coordinates.
-        elif (self.selectionMode == 0) and (self.freemove != FreeMode.NoFreeMove):
-            self.freemoved.emit(tuple((num, (
-                self.vpoints[num].cx,
-                self.vpoints[num].cy,
-                self.vpoints[num].angle,
-            )) for num in self.selections))
+        elif (self.select_mode == 0) and (self.freemove != FreeMode.NoFreeMove):
+            _emit_freemove(self, self.selections)
     self.selector.release()
     self.update()
     event.accept()
@@ -630,15 +673,23 @@ def mouseMoveEvent(self, event):
                     self.selector.sx / self.zoom,
                     self.selector.sy / -self.zoom,
                     len(selection),
-                    'link' if self.selectionMode == 1 else 'point'
+                    'point' if self.select_mode == 0 else
+                    'link' if self.select_mode == 1 else
+                    'solution'
                 ),
                 self
             )
-        elif self.selectionMode == 0:
+        elif self.select_mode == 0:
             if self.freemove == FreeMode.Translate:
                 #Free move translate function.
-                mouse_x = _snap(self, x - self.selector.x / self.zoom, False)
-                mouse_y = _snap(self, y - self.selector.y / -self.zoom, False)
+                mouse_x = _snap(self,
+                    x - self.selector.x / self.zoom,
+                    is_zoom=False
+                )
+                mouse_y = _snap(self,
+                    y - self.selector.y / -self.zoom,
+                    is_zoom=False
+                )
                 QToolTip.showText(
                     event.globalPos(),
                     "{:+.02f}, {:+.02f}".format(mouse_x, mouse_y),
@@ -646,10 +697,7 @@ def mouseMoveEvent(self, event):
                 )
                 for num in self.selections:
                     vpoint = self.vpoints[num]
-                    vpoint.move((
-                        mouse_x + vpoint.x,
-                        mouse_y + vpoint.y
-                    ))
+                    vpoint.move((mouse_x + vpoint.x, mouse_y + vpoint.y))
             elif self.freemove == FreeMode.Rotate:
                 #Free move rotate function.
                 alpha = atan2(y, x) - atan2(
@@ -664,10 +712,7 @@ def mouseMoveEvent(self, event):
                     vpoint = self.vpoints[num]
                     r = hypot(vpoint.x, vpoint.y)
                     beta = atan2(vpoint.y, vpoint.x)
-                    vpoint.move((
-                        r * cos(beta + alpha),
-                        r * sin(beta + alpha)
-                    ))
+                    vpoint.move((r * cos(beta + alpha), r * sin(beta + alpha)))
                     if vpoint.type != 0:
                         vpoint.rotate(self.vangles[num] + degrees(beta + alpha))
             elif self.freemove == FreeMode.Reflect:
@@ -684,10 +729,7 @@ def mouseMoveEvent(self, event):
                     if vpoint.type == 0:
                         vpoint.move((vpoint.x * fx, vpoint.y * fy))
                     else:
-                        vpoint.move(
-                            (vpoint.x * fx, vpoint.y * fy),
-                            (vpoint.x * fx, vpoint.y * fy)
-                        )
+                        vpoint.move((vpoint.x * fx, vpoint.y * fy))
                         if (x > 0) != (y > 0):
                             vpoint.rotate(180 - self.vangles[num])
         self.update()
@@ -699,25 +741,25 @@ def zoomToFit(self):
     """Zoom to fit function."""
     width = self.width()
     height = self.height()
-    width = width if not (width == 0) else 1
-    height = height if not (height == 0) else 1
+    width = width if width else 1
+    height = height if height else 1
     x_right, x_left, y_top, y_bottom = _zoomToFitLimit(self)
     inf = float('inf')
     if (inf in (x_right, y_bottom)) or (-inf in (x_left, y_top)):
         self.zoom_changed.emit(200)
-        self.ox = width/2
-        self.oy = height/2
+        self.ox = width / 2
+        self.oy = height / 2
         self.update()
         return
     x_diff = x_left - x_right
     y_diff = y_top - y_bottom
-    x_diff = x_diff if (x_diff != 0) else 1
-    y_diff = y_diff if (y_diff != 0) else 1
+    x_diff = x_diff if x_diff else 1
+    y_diff = y_diff if y_diff else 1
     if (width / x_diff) < (height / y_diff):
         factor = width / x_diff
     else:
         factor = height / y_diff
-    self.zoom_changed.emit(int(factor * self.marginFactor * 50))
-    self.ox = width / 2 - (x_left + x_right) / 2 *self.zoom
-    self.oy = height / 2 + (y_top + y_bottom) / 2 *self.zoom
+    self.zoom_changed.emit(int(factor * self.margin_factor * 50))
+    self.ox = (width - (x_left + x_right) * self.zoom) / 2
+    self.oy = (height + (y_top + y_bottom) * self.zoom) / 2
     self.update()